% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_synteny.R
\name{analysisEachCluster}
\alias{analysisEachCluster}
\title{Perform synteny analysis for identified clusters}
\usage{
analysisEachCluster(
  segmented_file,
  segmented_anchorpoints_file,
  genes_file,
  cluster_info_file,
  identified_cluster_file,
  hcheight = 0.3
)
}
\arguments{
\item{segmented_file}{The path to the segmented chromosome file.}

\item{segmented_anchorpoints_file}{The path to the segmented anchorpoints file.}

\item{genes_file}{genes.txt created by i-ADHoRe.}

\item{cluster_info_file}{The path to the clustering information file.}

\item{identified_cluster_file}{The path to the output file for identified clusters.}

\item{hcheight}{The cutoff height for cluster identification (default: 0.3).}
}
\value{
A list containing information about identified clusters and their p-values.
}
\description{
This function performs synteny analysis for clusters identified by hierarchical clustering.
}
