% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_synteny.R
\name{CalHomoConcentration}
\alias{CalHomoConcentration}
\title{Compute the -log10 of Poisson Distribution}
\usage{
CalHomoConcentration(m, n, q, k)
}
\arguments{
\item{m}{The total number of trials.}

\item{n}{The total number of possible outcomes.}

\item{q}{The observed number of successful outcomes.}

\item{k}{The expected number of successful outcomes.}
}
\value{
The -log10 of the p-value.
}
\description{
This function calculates the -log10 of the p-value of a Poisson distribution given the parameters.
}
