% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_synteny.R
\name{CalPvalue}
\alias{CalPvalue}
\title{Compute the P-value of a Cluster using the Poisson Distribution}
\usage{
CalPvalue(m, n, q, k)
}
\arguments{
\item{m}{The total number of all anchored points.}

\item{n}{The product of the remapped gene number of the query species and subject species.}

\item{q}{The number of anchored points in the cluster.}

\item{k}{The product of the remapped gene number of the segmented chromosomes
of the query species and subject species.}
}
\value{
The computed P-value.
}
\description{
This function computes the P-value of a cluster using the Poisson distribution.
}
