% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gff_from_file}
\alias{check_gff_from_file}
\title{Check the annotation file from a file location}
\usage{
check_gff_from_file(gff_input_name, gff_input_path)
}
\arguments{
\item{gff_input_name}{The prefix of the annotation file}

\item{gff_input_path}{The path of the annotation file}
}
\value{
A gff file path includes standardized annotation info.
}
\description{
Check the annotation file from a file location
}
\examples{
gff_temp <- check_gff_from_file(informal_name_temp, data_table[i, 3])
}
