% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gff_input}
\alias{check_gff_input}
\title{Check the uploading annotation file}
\usage{
check_gff_input(gff_input_name, gff_input_path)
}
\arguments{
\item{gff_input_name}{The prefix of the annotation file}

\item{gff_input_path}{The annotation file created by Shiny uploading function}
}
\value{
A gff file path includes standardized annotation info.
}
\description{
Check the uploading annotation file
}
\examples{
check_gff_input(informal_name, input[[paste0("gff_", 1)]])
}
