% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proteome_from_file}
\alias{check_proteome_from_file}
\title{Check the proteom file from the file location}
\usage{
check_proteome_from_file(proteome_name, proteome_input)
}
\arguments{
\item{proteome_name}{The prefix of the proteome file}

\item{proteome_input}{The path of the proteome file}
}
\value{
A proteome fasta file path includes standardized protein info.
}
\description{
Check the proteom file from the file location
}
\examples{
proteome_temp <- check_proteome_from_file(informal_name_temp, proteome)
}
