% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proteome_input}
\alias{check_proteome_input}
\title{Check the uploading proteome file}
\usage{
check_proteome_input(proteome_name, proteome_input)
}
\arguments{
\item{proteome_name}{The prefix of the proteome file}

\item{proteome_input}{The proteome file created by Shiny uploading function}
}
\value{
A proteome fasta file path includes standardized protein info.
}
\description{
Check the uploading proteome file
}
\examples{
proteome_temp <- check_proteome_input(informal_name_temp, input[[proteome]])
}
