% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_synteny.R
\name{cluster_synteny}
\alias{cluster_synteny}
\title{Cluster Synteny Data and Generate Trees}
\usage{
cluster_synteny(
  segmented_file,
  segmented_anchorpoints_file,
  genes_file,
  out_file
)
}
\arguments{
\item{segmented_file}{A character string specifying the file path for segmented data.}

\item{segmented_anchorpoints_file}{A character string specifying the file path for segmented anchorpoints.}

\item{genes_file}{A character string specifying the file path for genes information created by i-ADHoRe.}

\item{out_file}{A character string specifying the output file path for saving cluster information.}
}
\value{
NULL (output files are generated with the specified information).
}
\description{
This function clusters synteny data based on calculated p-values and generates trees
for both column-based and row-based clustering. It then saves the cluster information and
trees to output files.
}
