% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAnchorPointDepth.R
\name{computing_depth}
\alias{computing_depth}
\title{Compute the Depth of Anchored Points}
\usage{
computing_depth(
  anchorpoint_ks_file,
  multiplicon_id,
  selected_query_chr,
  selected_subject_chr = NULL
)
}
\arguments{
\item{anchorpoint_ks_file}{The file containing anchorpoint and Ks data.}

\item{multiplicon_id}{The ID of the multiplicon to consider.}

\item{selected_query_chr}{A list of selected query chromosomes.}

\item{selected_subject_chr}{A list of selected subject chromosomes (optional).}
}
\value{
A list containing depth data frames, including "query_depth" and "subject_depth" if subject chromosomes are specified, or "depth" if not.
}
\description{
This function calculates the depth of anchored points based on the provided parameters.
}
