% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAnchorPointDepth.R
\name{computing_depth_paranome}
\alias{computing_depth_paranome}
\title{Compute the Depth of Anchored Points in a Paranome Comparison}
\usage{
computing_depth_paranome(
  anchorpoint_ks_file,
  multiplicon_id,
  selected_query_chr
)
}
\arguments{
\item{anchorpoint_ks_file}{The file containing anchor point and Ks value data.}

\item{multiplicon_id}{The IDs of the multiplicons to consider.}

\item{selected_query_chr}{The list of selected query chromosomes.}
}
\value{
A list containing the depth dataframe.
}
\description{
This function computes the depth of anchored points in a paranome comparison based on the provided parameters.
}
