% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{convert_wgd2kevins}
\alias{convert_wgd2kevins}
\title{Convert}
\usage{
convert_wgd2kevins(file, species_name)
}
\arguments{
\item{file}{input file}

\item{species_name}{species name}
}
\description{
Convert
}
\examples{
convert_wgd2kevins(file, "rice")
}
