% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountOrthologs.R
\name{CountOrthologs}
\alias{CountOrthologs}
\title{Count Ortholog Genes in a Species}
\usage{
CountOrthologs(atomic.df, species)
}
\arguments{
\item{atomic.df}{A data frame containing information about ortholog genes.
It should have the following columns:
- multiplicon: The multiplicon identifier.
- geneX: The gene identifier in speciesX.
- speciesX: The species name for geneX.
- listX: The chromosome or list identifier for geneX.
- coordX: The coordinate information for geneX.
- geneY: The gene identifier in speciesY.
- speciesY: The species name for geneY.
- listY: The chromosome or list identifier for geneY.
- coordY: The coordinate information for geneY.
- level: The orthology level.
- num_anchors: The number of anchors.
- is_real: A flag indicating if the data is real.
- Ks: The Ks value.}

\item{species}{The species for which ortholog gene counts should be computed.}
}
\value{
A data frame summarizing the counts of ortholog genes for each chromosome.
}
\description{
This function counts ortholog genes in a given species based on input data.
}
\examples{
# Example usage:
# Create a sample atomic.df data frame
atomic.df <- data.frame(
  multiplicon=c(2, 2, 2),
  geneX=c("GSVIVT01016360001", "GSVIVT01016362001", "GSVIVT01016362001"),
  speciesX=c("Vitis_vinifera", "Vitis_vinifera", "Vitis_vinifera"),
  listX=c("chr13:181-369", "chr13:181-369", "chr13:181-369"),
  coordX=c(188, 187, 1875),
  geneY=c("Os01t0854500-01", "Os01t0854000-01", "Os05t0449200-01"),
  speciesY=c("Oryza_sativa", "Oryza_sativa", "Oryza_sativa"),
  listY=c("chr01:3463-3614", "chr01:3463-3614", "chr05:1610-1714"),
  coordY=c(135, 133, 14),
  level=c(10000, 10000, 10000),
  num_anchors=c(10000, 10000, 10000),
  is_real=c(-1, -1, -1),
  Ks=c(81.3724, 77.5539, 76.6121)
)

# Calculate ortholog gene counts for "Vitis_vinifera"
ortholog_counts <- CountOrthologs(atomic.df, species="Vitis_vinifera")
ortholog_counts
}
