% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_cmd}
\alias{create_ksrates_cmd}
\title{Create ksrates command Shell file}
\usage{
create_ksrates_cmd(input, ksratesconf, cmd_file)
}
\arguments{
\item{input}{The Input object of Shiny.}

\item{ksratesconf}{The file of ksrates configure infomation.}

\item{cmd_file}{The shell file is used to save the commmand line of ksrates.}
}
\description{
Create ksrates command Shell file
}
\examples{
create_ksrates_cmd(input, "ksrates_conf.txt", ksrates_cmd_sh_file)
}
