% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_cmd_from_table}
\alias{create_ksrates_cmd_from_table}
\title{Create ksrates command Shell file}
\usage{
create_ksrates_cmd_from_table(
  data_table,
  ksratesconf,
  cmd_file,
  wgd_cmd_file,
  focal_species
)
}
\arguments{
\item{data_table}{A data frame includes three columns: species name, proteome file, annotation file.}

\item{ksratesconf}{The file of ksrates configure infomation.}

\item{cmd_file}{The shell file is used to save the commmand line of ksrates.}

\item{wgd_cmd_file}{The shell file is used to save the command line of wgd.}

\item{focal_species}{The name of the focal species used in ksrates.}
}
\description{
Create ksrates command Shell file
}
\examples{
 create_ksrates_cmd_from_table(data_table, "ksrates_conf.txt", ksrates_cmd_sh_file, wgd_cmd_sh_file, input$select_focal_species)
}
