% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_configure_file_based_on_table}
\alias{create_ksrates_configure_file_based_on_table}
\title{Create ksrates configure file based on the data table}
\usage{
create_ksrates_configure_file_based_on_table(
  data_table,
  focal_species,
  newick_tree_file,
  ksrates_conf_file,
  species_info_file
)
}
\arguments{
\item{data_table}{A data frame includes three columns: species name, proteome file, annotation file.}

\item{focal_species}{The name of the focal species used in ksrates}

\item{newick_tree_file}{The path of newick tree file}

\item{ksrates_conf_file}{The file is used to store the configure of ksrates.}

\item{species_info_file}{The file is used to store the updated info.}
}
\description{
Create ksrates configure file based on the data table
}
\examples{
create_ksrates_configure_file_based_on_table(
 data_table,
 input$select_focal_species,
 input$newick_tree,
 ksratesconf,
 speciesinfoconf
)
}
