% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_configure_file_v2}
\alias{create_ksrates_configure_file_v2}
\title{Creating ksrates configure file}
\usage{
create_ksrates_configure_file_v2(input, ksrates_conf_file, species_info_file)
}
\arguments{
\item{input}{The Input object of Shiny.}

\item{ksrates_conf_file}{The file is used to store the configure of ksrates.}

\item{species_info_file}{The file is used to store the updated info.}
}
\description{
Creating ksrates configure file
}
\examples{
create_ksrates_configure_file_v2(input, ksratesconf, speciesinfoconf)
}
