% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_features.R
\name{dfltCounts}
\alias{dfltCounts}
\title{dfltCounts}
\usage{
dfltCounts(
  x,
  gridsize = rep(64, NCOL(x)),
  h = rep(0, NCOL(x)),
  supp = 3.7,
  range.x,
  w
)
}
\arguments{
\item{x}{The input data, which should be a numeric matrix.}

\item{gridsize}{A vector specifying the number of bins along each dimension.}

\item{h}{A vector specifying the bandwidth (smoothing parameter) along each dimension.}

\item{supp}{A parameter for determining the range of the bins.}

\item{range.x}{A list specifying the range of values for each dimension.}

\item{w}{A vector of weights for the data points.}
}
\value{
A list containing the binned counts and the range of values for each dimension.
}
\description{
This function bins the input data into a regular grid.
}
