% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_features.R
\name{drvkde}
\alias{drvkde}
\title{drvkde}
\usage{
drvkde(x, drv, bandwidth, gridsize, range.x, binned = FALSE, se = TRUE, w)
}
\arguments{
\item{x}{The input data.}

\item{drv}{The order of the derivative to compute.}

\item{bandwidth}{The bandwidth (smoothing parameter) along each dimension.}

\item{gridsize}{The size of the grid.}

\item{range.x}{A list specifying the range of values for each dimension.}

\item{binned}{A logical indicating whether the input data is already binned.}

\item{se}{A logical indicating whether to compute standard errors.}

\item{w}{A vector of weights for the data points.}
}
\value{
A list containing the estimated density or derivative, and optionally, standard errors.
}
\description{
Compute the mth derivative of a binned d-variate kernel density estimate based on grid counts.
}
