% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_first_part}
\alias{extract_first_part}
\title{Extract the first part of a string by splitting it at tab characters.}
\usage{
extract_first_part(name)
}
\arguments{
\item{name}{The input string to be split.}
}
\value{
Returns the first part of the input string.
}
\description{
This function takes a string and splits it at tab characters. It then
returns the first part of the resulting character vector.
}
