% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_synteny.R
\name{extractCluster}
\alias{extractCluster}
\title{Extract clusters based on specified scaffolds}
\usage{
extractCluster(segs.df, atomic.df, scaf.bycol, scaf.byrow)
}
\arguments{
\item{segs.df}{A data frame containing segment information.}

\item{atomic.df}{A data frame containing atomic anchorpoints.}

\item{scaf.bycol}{A character vector specifying scaffolds for the query species.}

\item{scaf.byrow}{A character vector specifying scaffolds for the subject species.}
}
\value{
A list containing two data frames: "segs" for segment information and "atomic" for atomic anchorpoints.
}
\description{
This function extracts clusters based on the specified scaffolds for both query and subject species.
It filters the data frames containing segment information and atomic anchorpoints to retain only the relevant clusters.
}
