% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_resample_95_CI.R
\name{find_peaks}
\alias{find_peaks}
\title{Find Peaks in a Numeric Vector}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{A numeric vector in which peaks will be identified.}

\item{m}{An integer indicating the half-width of the neighborhood to consider when identifying peaks. A larger value of \code{m} makes peak detection less sensitive.}
}
\value{
A numeric vector containing the indices of the identified peaks in the input vector \code{x}.
}
\description{
This function identifies peaks in a numeric vector by analyzing the shape of the curve.
}
\examples{
# Generate some example data
x <- c(1, 3, 7, 2, 6, 8, 5, 4, 9, 3, 2, 1)

# Find peaks in the data with a half-width of 2
peaks <- find_peaks(x, m=2)

# Print the indices of the identified peaks
print(peaks)
}
