% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{generate_ksd}
\alias{generate_ksd}
\title{Convert}
\usage{
generate_ksd(ks_df, bin_width = 0.01, maxK = 5)
}
\arguments{
\item{ks_df}{A data frame including Ks and Weights.}

\item{bin_width}{Bin width for generating the distribution, default 0.01.}

\item{maxK}{Maximum Ks value for generating the distribution, default 5.}
}
\value{
A list including:
*\code{Ks}: the Ks distribution;
*\code{bin_width}: bin width;
*\code{maxK}: maximum Ks.
}
\description{
Convert
}
\examples{
generate_ksd(ks_df, bin_width=0.01)
}
