% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.not.null}
\alias{is.not.null}
\title{Check if an Object is Not NULL}
\usage{
is.not.null(x)
}
\arguments{
\item{x}{An R object to check.}
}
\value{
A logical value indicating whether the object is not NULL.
}
\description{
This function checks if an object is not NULL.
}
