% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_fasta_cds}
\alias{is_fasta_cds}
\title{Check if a file is in FASTA format with cds sequences.}
\usage{
is_fasta_cds(file_path)
}
\arguments{
\item{file_path}{The path to the input file.}
}
\value{
TRUE if the file is in FASTA format with cds sequences, FALSE otherwise.
}
\description{
This function checks whether a given file is in FASTA format with cds sequences.
}
