% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_header_contains_pipe}
\alias{is_header_contains_pipe}
\title{Check if the header of sequences contains a pipe character ("|").}
\usage{
is_header_contains_pipe(file_path)
}
\arguments{
\item{file_path}{The path to the file to be checked.}
}
\value{
Returns TRUE if the header contains a pipe character, otherwise FALSE.
}
\description{
This function reads the first line of a fasta file and checks if the header contains the
pipe character "|".
}
