% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/running_emmix.R
\name{ks_mclust_v2}
\alias{ks_mclust_v2}
\title{ks_mclust_v2}
\usage{
ks_mclust_v2(input_data)
}
\arguments{
\item{input_data}{The input data for clustering and modeling.}
}
\value{
A data frame containing clustering and modeling results.
}
\description{
A wrapper to run emmix modeling using the mclust package.
}
