% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_informal_name_to_latin_name}
\alias{map_informal_name_to_latin_name}
\title{Remapping the temporary informal name to the Latin name}
\usage{
map_informal_name_to_latin_name(sp_gff_info_xls)
}
\arguments{
\item{sp_gff_info_xls}{A file includes the Latin name, informal name, and gff file}
}
\value{
A data frame includes the corresponding pair of Latin and informal name
}
\description{
Remapping the temporary informal name to the Latin name
}
\examples{
names_df <- map_informal_name_to_latin_name(species_info_file[1])
}
