% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmix.R
\name{mix_logNormal_Ks}
\alias{mix_logNormal_Ks}
\title{Log-Normal mixturing analyses of a Ks distributions for the whole paranome}
\usage{
mix_logNormal_Ks(ksv, G = 1:5, k.nstart = 500, maxK = 5)
}
\arguments{
\item{ksv}{A \code{ksv} object.}

\item{G}{An integer vector specifying the range of the mixtured components.
A BIC is calculated for each component. The default is G=1:5.
For a formal analysis, it is recommended to use 1:10.}

\item{k.nstart}{How many random sets should be chosen in the k-means
clustering.
For a formal analysis, it is recommended to use 500.}

\item{maxK}{Maximum Ks values used in the mixture modeling analysis.}
}
\value{
A data frame with seven variables.
}
\description{
Log-Normal mixturing analyses of a Ks distributions for the whole paranome
}
