% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_coordinates_for_segments.v2.R
\name{obtain_coordiantes_for_segments}
\alias{obtain_coordiantes_for_segments}
\title{Obtain coordinates for segments in a comparison}
\usage{
obtain_coordiantes_for_segments(
  seg_file,
  sp1,
  gff_file1,
  out_file,
  sp2 = NULL,
  gff_file2 = NULL
)
}
\arguments{
\item{seg_file}{The file containing segment data.}

\item{sp1}{The species name for the first genome.}

\item{gff_file1}{The GFF file for the first genome.}

\item{out_file}{The output file to store the merged position data.}

\item{sp2}{The species name for the second genome (optional).}

\item{gff_file2}{The GFF file for the second genome (optional).}
}
\value{
NULL (the results are saved in the output file).
}
\description{
This function retrieves the coordinates for segments in a comparison based on the provided parameters.
}
