% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmix.R
\name{parse_EMMIX}
\alias{parse_EMMIX}
\title{Read the EMMIX output for a range of components}
\usage{
parse_EMMIX(emmix.out, G = 1:3)
}
\arguments{
\item{emmix.out}{The output file from EMMIX software.}

\item{G}{An integer vector specifying the range of the mixture components.
The default is G=1:3.}
}
\value{
A data frame with seven variables.
}
\description{
Read the EMMIX output for a range of components
}
