% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmix.R
\name{parse_one_EMMIX}
\alias{parse_one_EMMIX}
\title{Read the EMMIX output for a specify number of components}
\usage{
parse_one_EMMIX(emmix.out, ncomponent = 3)
}
\arguments{
\item{emmix.out}{The output file from EMMIX software.}

\item{ncomponent}{Number of components to read from the file.}
}
\value{
A data frame with seven variables.
}
\description{
Read the EMMIX output for a specify number of components
}
