% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{plot.ksv}
\alias{plot.ksv}
\title{Draw Ks distribution using output from \code{wgd}}
\usage{
\method{plot}{ksv}(
  ksv,
  bin_width = 0.1,
  maxK = 5,
  maxY = 1000,
  plot_mode = "weighted",
  color = "gray",
  pdf_file = NULL,
  ...
)
}
\arguments{
\item{bin_width}{bin width of the histgram, default 0.1, minimum 0.01}

\item{maxK}{maximum Ks value, default 5}

\item{maxY}{maximum value for y-axis, default 1000}

\item{plot_mode}{modes to account for redundant duplicate, including \code{weighted} (default), \code{average}, \code{min}, and \code{pairwise}}

\item{color}{color of the histgram, default "darkgray"}

\item{pdf_file}{plot the figure in a pdf file with a specific name or not \code{NULL} (default)}

\item{file}{\code{ksd} output file from \code{wgd}, usually with a name \verb{*.ks.tsv}}

\item{include.outliers}{wheter to include outliers in the \verb{wgd ksd} outpuf, default \code{FALSE}}

\item{minK}{minimum Ks value, default 0}

\item{minAlnLen}{minimum alignment length, default 0}

\item{minIdn}{minimum alignment identity, default 0}

\item{minCov}{minimum alignment coverage, default 0}
}
\description{
Draw Ks distribution using output from \code{wgd}
}
\examples{
ksv <- read.wgd_ksd(file)
plot.ksv(ksv)
}
