% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmix.R
\name{plot_Ks_mix}
\alias{plot_Ks_mix}
\title{Draw a Ks distribution with fitted Gaussian}
\usage{
plot_Ks_mix(
  ksv,
  ncomponent = 3,
  EMMIX.file = NULL,
  bin_width = 0.1,
  maxK = 5,
  maxY = 1000,
  ...
)
}
\arguments{
\item{ksv}{A \code{ksv} object}

\item{ncomponent}{Number of components, default 3.}

\item{EMMIX.file}{The output file of EMMIX software.}

\item{bin_width}{Bin width for the Ks distribution, default 0.1}

\item{maxK}{Maximum Ks value, default 5.}

\item{maxY}{Maximum number of y-axis, default 1000.}

\item{...}{other parameters in plot.ksd or plot.ksv and mix_logNormal_Ks}
}
\description{
Draw a Ks distribution with fitted Gaussian
}
\examples{
plot_Ks_mix(ksv)
}
