% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{plot_ksv_density}
\alias{plot_ksv_density}
\title{Draw a density plot of ksv}
\usage{
plot_ksv_density(mids, counts, maxK, spar = 0.2, color = "#424242")
}
\arguments{
\item{mids}{the data of ksv}

\item{counts}{counts}

\item{maxK}{maximum Ks}

\item{spar}{the smooth level}

\item{color}{color}
}
\description{
Draw a density plot of ksv
}
\examples{
plot_ksv_density(mids, count, 5)
}
