% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{read.dating_anchors}
\alias{read.dating_anchors}
\title{Read the output of Ks values for anchor pairs from the dating pipeline}
\usage{
read.dating_anchors(file, ks_col = 9, header = TRUE)
}
\arguments{
\item{file}{Output file of Ks values for anchor pairs from the dating
pipeline. If it a tabular file with Ks values not from the dating pipeline,
please denote the column of Ks values using \code{ks_col}}

\item{ks_col}{Index of column for Ks values, default 9.}

\item{header}{Wheather the file has a header or not, default TRUE.}
}
\value{
A \code{ksv} object, which is a list including:
\itemize{
\item \code{ks_df}: the data frame that used for following analysis
\item \code{ks_dist}: a list including a vector of Ks values in the distribution
\item \code{raw_df}: raw data
\item \code{filters}: filters that applied to the raw data
}
}
\description{
Read the output of Ks values for anchor pairs from the dating pipeline
}
\examples{
ksv <- read.dating_anchors(file)
}
\seealso{
\code{generate_ksd}
}
