% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data_file}
\alias{read_data_file}
\title{Read a upload table file}
\usage{
read_data_file(uploadfile)
}
\arguments{
\item{uploadfile}{The object of the file through uploading function of Shiny.}
}
\value{
A data frame includes the data from the file.
}
\description{
Read a upload table file
}
\examples{
read_data_file(input$upload_data_file)[["V1"]]
}
