% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksv.R
\name{read_KTpipeline}
\alias{read_KTpipeline}
\title{Read file from KT pipeline}
\usage{
read_KTpipeline(file)
}
\arguments{
\item{file}{Output from KT pipeline}
}
\description{
Read file from KT pipeline
}
\examples{
read_KTpipeline(file)
}
