% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitution_rate_correction.R
\name{relativeRate}
\alias{relativeRate}
\title{relativeRate}
\usage{
relativeRate(
  ksv2out_1_file,
  ksv2out_2_file,
  ksv_between_file,
  KsMax,
  low = 0.025,
  up = 0.975,
  bs = 1000
)
}
\arguments{
\item{ksv2out_1_file}{A character string specifying the path to the first input data file.}

\item{ksv2out_2_file}{A character string specifying the path to the second input data file.}

\item{ksv_between_file}{A character string specifying the path to the third input data file.}

\item{KsMax}{A numeric value representing a maximum threshold for Ks values.}

\item{low}{A numeric value specifying the lower quantile for bootstrapping. Default is 0.025.}

\item{up}{A numeric value specifying the upper quantile for bootstrapping. Default is 0.975.}

\item{bs}{An integer specifying the number of bootstrap iterations. Default is 1000.}
}
\value{
A list containing computed relative rates and their confidence intervals.
}
\description{
Compute relative rates using input data files and statistical computations.
}
