% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_inner_stop_codon_sequence}
\alias{remove_inner_stop_codon_sequence}
\title{Remove Genes Contain Stop Codons within the Sequence}
\usage{
remove_inner_stop_codon_sequence(sequence)
}
\arguments{
\item{sequence}{A nucleotide sequence as a character string.}
}
\value{
A character string or NULL.
}
\description{
This function removes the gene contains stop codons (TAA, TAG, TGA, taa, tag, tga)
within its sequence.
}
