% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_old_directories.R
\name{remove_old_dirs}
\alias{remove_old_dirs}
\title{Remove directories older than a specified day}
\usage{
remove_old_dirs(
  base_dir,
  max_age_in_days = 3,
  log_file = "remove_old_dirs.log",
  verbose = FALSE
)
}
\arguments{
\item{base_dir}{The base directory to search for old directories.}

\item{max_age_in_days}{The maximum age (in days) for directories to be considered old.}

\item{log_file}{The name of the log file to store information about removed directories and errors.}

\item{verbose}{A logical value indicating whether to print messages to the console.}
}
\value{
The function does not return anything. It logs information about removed directories and errors.
}
\description{
This function removes directories in the specified base directory that are
older than a specified maximum age in days. It logs the removed directories
and any errors encountered during removal.
}
