% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_informal_name_to_latin_name}
\alias{replace_informal_name_to_latin_name}
\title{Replace the infomal name to Latin name}
\usage{
replace_informal_name_to_latin_name(names_df, input)
}
\arguments{
\item{names_df}{The data frame includes the corresponding pair of Latin and informal name}

\item{input}{The informal name}
}
\value{
A string includes the Latin name
}
\description{
Replace the infomal name to Latin name
}
\examples{
replace_informal_name_to_latin_name(names_df, x)
}
