% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmix.R
\name{run_emmix_kmeas}
\alias{run_emmix_kmeas}
\title{A wrapper to run EM analysis of \(ln\) Ks values with k-means}
\usage{
run_emmix_kmeas(v, k.centers = 2, k.nstart = 500)
}
\arguments{
\item{v}{A list include a vector of Ks values namely \code{ks_value},
and a bolean variable namely \code{log}}

\item{k.centers}{Number of k-means centers, default 2.}

\item{k.nstart}{Number of random start of k-means clustering, default 10.
For a formal analysis, it is recommended to use 500.}
}
\value{
A list, i.e., the original output of mclust::emV
}
\description{
A wrapper to run EM analysis of \(ln\) Ks values with k-means
}
