% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_features.R
\name{symconv.ks}
\alias{symconv.ks}
\title{symconv.ks}
\usage{
symconv.ks(rr, ss, skewflag)
}
\arguments{
\item{rr}{The first input vector.}

\item{ss}{The second input vector.}

\item{skewflag}{A scalar value to apply skew correction.}
}
\value{
A vector representing the result of the symmetric convolution.
}
\description{
Perform symmetric convolution using FFT.
}
