% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_features.R
\name{symconv3D.ks}
\alias{symconv3D.ks}
\title{symconv3D.ks}
\usage{
symconv3D.ks(rr, ss, skewflag = rep(1, 3))
}
\arguments{
\item{rr}{The first input 3D array.}

\item{ss}{The second input 3D array.}

\item{skewflag}{A vector of three scalar values for skew correction along each dimension.}
}
\value{
A 3D array representing the result of the symmetric 3D convolution.
}
\description{
Perform symmetric 3D convolution using FFT.
}
