% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_features.R
\name{symconv4D.ks}
\alias{symconv4D.ks}
\title{symconv4D.ks}
\usage{
symconv4D.ks(rr, ss, skewflag = rep(1, 4), fftflag = rep(TRUE, 2))
}
\arguments{
\item{rr}{The first input 4D array.}

\item{ss}{The second input 4D array.}

\item{skewflag}{A vector of four scalar values for skew correction along each dimension.}

\item{fftflag}{A vector of two Boolean values for FFT flag.}
}
\value{
A 4D array representing the result of the symmetric 4D convolution.
}
\description{
Perform symmetric 4D convolution using FFT.
}
