% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_tree_from_timetree.R
\name{TimeTreeFecher}
\alias{TimeTreeFecher}
\title{Extracts a timetree from TimeTree.org based on species names.}
\usage{
TimeTreeFecher(input_file, prefix)
}
\arguments{
\item{input_file}{A character string specifying the path to the file containing species names.}

\item{prefix}{A character string providing the prefix for the output file.}
}
\value{
A timetree object representing the estimated divergence times between species.
}
\description{
This function takes a file with species names as input and a prefix to define the output.
}
