% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gff_from_file}
\alias{check_gff_from_file}
\title{Check and Process GFF Input File from a Specific Path}
\usage{
check_gff_from_file(gff_input_name, gff_input_path, working_wd)
}
\arguments{
\item{gff_input_name}{The informal name of the GFF input file.}

\item{gff_input_path}{The path to the GFF input file.}

\item{working_wd}{A character string specifying the working directory to be used.}
}
\value{
A string containing the processed GFF file's path.
}
\description{
This function checks the type of GFF input file specified by its path and processes it accordingly.
}
