% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_gff_input}
\alias{check_gff_input}
\title{Check and Prepare GFF/GTF Input File}
\usage{
check_gff_input(gff_input_name, gff_input_path, working_wd)
}
\arguments{
\item{gff_input_name}{A descriptive name for the GFF/GTF file.}

\item{gff_input_path}{The file path to the GFF/GTF file.}

\item{working_wd}{A character string specifying the working directory to be used.}
}
\value{
The path to the prepared GFF file for analysis.
}
\description{
This function checks the file format of a GFF/GTF input file and prepares it for analysis. It can handle both uncompressed and compressed formats.
}
