% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_proteome_input}
\alias{check_proteome_input}
\title{Check and Process Proteome Input File}
\usage{
check_proteome_input(proteome_name, proteome_input, working_wd)
}
\arguments{
\item{proteome_name}{The informal name of the proteome input file.}

\item{proteome_input}{The proteome input data.}

\item{working_wd}{A character string specifying the working directory to be used.}
}
\value{
A string containing the processed proteome file's path.
}
\description{
This function checks the type of proteome input file and processes it accordingly.
}
