% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkFileExistence}
\alias{checkFileExistence}
\title{Check File Existence in a Data Table}
\usage{
checkFileExistence(data_table, working_wd)
}
\arguments{
\item{data_table}{A data table with file paths in columns V2 and V3.}

\item{working_wd}{A path of the working directory}
}
\value{
This function has no return value. It prints messages to the console.
}
\description{
This function checks the existence of files specified in a data table.
}
