% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountOrthologs.R
\name{CountOrthologs}
\alias{CountOrthologs}
\title{Count Ortholog Genes in a Species}
\usage{
CountOrthologs(atomic.df, species)
}
\arguments{
\item{atomic.df}{A data frame containing information about ortholog genes.}

\item{species}{The species for which ortholog gene counts should be computed.}
}
\value{
A data frame summarizing the counts of ortholog genes for each chromosome.
}
\description{
This function counts ortholog genes in a given species based on input data.
}
\examples{
# Example usage:
ortholog_counts <- CountOrthologs(atomic.df, species="SpeciesA")
print(ortholog_counts)
}
