% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_cmd}
\alias{create_ksrates_cmd}
\title{Create Ksrates Command Files from Shiny Input}
\usage{
create_ksrates_cmd(input, ksratesconf, cmd_file)
}
\arguments{
\item{input}{The Input object of Shiny.}

\item{ksratesconf}{The path to the Ksrates configuration file.}

\item{cmd_file}{The path to the main Ksrates command file to be generated.}
}
\description{
Create Ksrates Command Files from Shiny Input
}
