% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_cmd_from_table}
\alias{create_ksrates_cmd_from_table}
\title{Create Ksrates Command Files from Data Table}
\usage{
create_ksrates_cmd_from_table(data_table, ksratesconf, cmd_file, focal_species)
}
\arguments{
\item{data_table}{The data table containing information about species.}

\item{ksratesconf}{The path to the Ksrates configuration file.}

\item{cmd_file}{The path to the main Ksrates command file to be generated.}

\item{focal_species}{The name of the focal species.}
}
\description{
This function generates command files for running Ksrates and related analyses based on a data table and configuration file.
}
