% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_configure_file_based_on_table}
\alias{create_ksrates_configure_file_based_on_table}
\title{Create Ksrates Configuration File Based on Data Table}
\usage{
create_ksrates_configure_file_based_on_table(
  data_table,
  focal_species,
  newick_tree_file,
  ksrates_conf_file,
  species_info_file,
  working_wd
)
}
\arguments{
\item{data_table}{The data table containing information about species, proteomes, and GFF files.}

\item{focal_species}{The name of the focal species.}

\item{newick_tree_file}{The path to the Newick tree file.}

\item{ksrates_conf_file}{The path to the Ksrates configuration file to be generated.}

\item{species_info_file}{The path to the species information file.}

\item{working_wd}{A character string specifying the working directory to be used.}
}
\description{
This function generates a Ksrates configuration file based on a data table and other parameters.
}
