% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_configure_file_v2}
\alias{create_ksrates_configure_file_v2}
\title{Create Ksrates Configuration File}
\usage{
create_ksrates_configure_file_v2(input, ksrates_conf_file, species_info_file)
}
\arguments{
\item{input}{The Input object of Shiny.}

\item{ksrates_conf_file}{The path to the Ksrates configuration file.}

\item{species_info_file}{The path to the species information file.}
}
\description{
This function generates a configuration file for the Ksrates pipeline based on Shiny input.
}
