% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_ksrates_expert_parameter_file}
\alias{create_ksrates_expert_parameter_file}
\title{Create ksrates Expert Parameter File}
\usage{
create_ksrates_expert_parameter_file(ksrates_expert_parameter_file)
}
\arguments{
\item{ksrates_expert_parameter_file}{The file is used to store the ksrates expert parameter}
}
\description{
Create ksrates Expert Parameter File
}
