% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_resample_95_CI.R
\name{generateKsDistribution}
\alias{generateKsDistribution}
\title{Generate the Ks Distribution}
\usage{
generateKsDistribution(ksraw, speciesName = NULL, maxK = 5)
}
\arguments{
\item{ksraw}{A numeric vector containing raw Ks values.}

\item{speciesName}{(Optional) A character string specifying the species name associated with the Ks values.}

\item{maxK}{A numeric value indicating the maximum Ks value to consider in the distribution.}
}
\value{
A numeric vector containing the binned Ks distribution.
}
\description{
This function generates a Ks (synonymous substitution rates) distribution from raw Ks values.
}
