% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_informal_name_to_latin_name}
\alias{map_informal_name_to_latin_name}
\title{Map Informal Names to Latin Names}
\usage{
map_informal_name_to_latin_name(sp_gff_info_xls)
}
\arguments{
\item{sp_gff_info_xls}{The path to the Excel file containing species information.}
}
\value{
A data frame with "latin_name" and "informal_name" columns.
}
\description{
This function reads information from an Excel file (XLS) containing columns "latin_name," "informal_name," and "gff." It extracts the "latin_name" and "informal_name" columns, performs some data manipulation, and returns a data frame with these two columns.
}
