% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_chromosome_length.R
\name{obtain_chromosome_length}
\alias{obtain_chromosome_length}
\title{obtain_chromosome_length}
\usage{
obtain_chromosome_length(species_info_file)
}
\arguments{
\item{species_info_file}{A character string specifying the path to the species information file.}
}
\value{
A list containing two data frames: len_df for chromosome lengths and num_df for mRNA counts.
}
\description{
Process species information file and extract chromosome lengths and mRNA counts from GFF files.
}
