% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain_coordinates_for_anchorpoints_ks.R
\name{obtain_coordiantes_for_anchorpoints_ks}
\alias{obtain_coordiantes_for_anchorpoints_ks}
\title{Obtain Coordinates and Ks Values for Anchorpoints}
\usage{
obtain_coordiantes_for_anchorpoints_ks(
  anchorpoints,
  anchorpoints_ks,
  genes_file,
  out_file,
  out_ks_file,
  species
)
}
\arguments{
\item{anchorpoints}{A character string specifying the file path for anchorpoints data.}

\item{anchorpoints_ks}{A character string specifying the file path for anchorpoints Ks values data.}

\item{genes_file}{A character string specifying the file path for genes information.}

\item{out_file}{A character string specifying the output file path for coordinates.}

\item{out_ks_file}{A character string specifying the output file path for Ks values.}

\item{species}{A character string specifying the species name.}
}
\value{
NULL (output files are generated with the specified information).
}
\description{
This function extracts coordinates and Ks (synonymous substitution rate) values for anchorpoints
from input data and merges them into a single output file.
}
