% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/obtain_coordinates_for_segments.multiple_species.R
\name{obtain_coordinates_for_segments_multiple}
\alias{obtain_coordinates_for_segments_multiple}
\title{Obtain Coordinates for Segments in Multiple Synteny Blocks}
\usage{
obtain_coordinates_for_segments_multiple(seg_df, gff_df, input, out_file)
}
\arguments{
\item{seg_df}{A dataframe containing information about synteny segments.}

\item{gff_df}{A dataframe containing GFF (General Feature Format) information.}

\item{input}{A list containing input data, typically multiple synteny query chromosomes.}

\item{out_file}{A character string specifying the output file path.}
}
\value{
A dataframe with coordinates for segments within multiple synteny blocks.
}
\description{
This function extracts coordinates for segments within multiple synteny blocks based on input dataframes.
}
