% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_data_file}
\alias{read_data_file}
\title{Read Data from Uploaded File}
\usage{
read_data_file(uploadfile)
}
\arguments{
\item{uploadfile}{The object representing the uploaded file obtained through the Shiny upload function.}
}
\value{
A data frame containing the data from the uploaded file.
}
\description{
This function reads data from an uploaded file in a Shiny application and returns it as a data frame.
}
