% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_informal_name_to_latin_name}
\alias{replace_informal_name_to_latin_name}
\title{Replace Informal Names with Latin Names}
\usage{
replace_informal_name_to_latin_name(names_df, input)
}
\arguments{
\item{names_df}{A data frame with "latin_name" and "informal_name" columns.}

\item{input}{The input string that may contain informal species names.}
}
\value{
A modified input string with informal names replaced by Latin names.
}
\description{
This function takes a data frame \code{names_df} containing "latin_name" and "informal_name" columns and an \code{input} string as input. It replaces informal species names in the \code{input} string with their corresponding Latin names based on the information in \code{names_df}. If the \code{input} string contains underscores ("_"), it assumes a comparison between two species and replaces both informal names. Otherwise, it replaces the informal name in the \code{input} string.
}
