% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_resample_95_CI.R
\name{resampleKsDistribution}
\alias{resampleKsDistribution}
\title{Resample a Ks Distribution}
\usage{
resampleKsDistribution(ks, maxK = 5)
}
\arguments{
\item{ks}{A numeric vector representing the Ks distribution to be resampled.}

\item{maxK}{A numeric value indicating the maximum Ks value to consider in the distribution.}
}
\value{
A numeric vector containing a resampled Ks distribution.
}
\description{
This function resamples a given Ks (synonymous substitution rates) distribution.
}
