% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiZer.R
\name{SiZer}
\alias{SiZer}
\title{SiZer (Significant Zero Crossings)}
\usage{
SiZer(x, bw, gridsize, signifLevel = 0.05)
}
\arguments{
\item{x}{A numeric vector containing the data for which you want to calculate SiZer.}

\item{bw}{Bandwidth parameter for kernel density estimation. If not provided, default values are used.}

\item{gridsize}{A vector specifying the grid size for SiZer. Default is c(401, 151).}

\item{signifLevel}{The significance level for SiZer. Default is 0.05.}
}
\value{
A list containing SiZer results, including the SiZer curve, the SiZer map, and the bandwidth.
}
\description{
The SiZer (Significant Zero Crossings) method is a technique used for assessing the statistical significance of zero crossings in data density estimation.
}
